Add support for extracting comments that are on the same line as the field
And use that as description for the field (shown in Wireshark's status bar)

It would be interesting to see if we should do the same for comments above
the field declaration. Seems like a good addition for completeness' sake,
but also Wireshark's presentation isn't very friendly to long, potentially
multi-line comments.